Fit test reorg style better in workgroup tests
